# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coding Platform API",
    "description": "REST API for the Coding Platform — an online judge that allows candidates to write, run, and submit code solutions against automated test suites.\n\n**Key capabilities:**\n- Run code against a single input (experimentation mode)\n- Submit a full solution for judging against visible + hidden tests\n",
    "contact": {
      "name": "Coding Platform",
      "url": "https://github.com/Sanjeev1393/Coding-Platform"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://your-service.onrender.com",
      "description": "Production server (Render)"
    }
  ],
...
  "paths": {
...
    "/api/v1/questions/{id}": {
      "get": {
        "tags": [
          "Questions API"
        ],
        "summary": "Get question by ID",
        "description": "Returns full problem details for a single question — including description, function signature, starter codes, and visible test cases. Returns 404 if the question ID does not exist.",
        "operationId": "getQuestionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Question identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "two-sum"
          }
        ],
        "responses": {
          "200": {
            "description": "Question found and returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "404": {
            "description": "No question found with the given ID"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FunctionParam": {
        "type": "object",
        "description": "A single named parameter of a function signature",
        "properties": {
          "name": {
            "type": "string",
            "description": "Parameter name",
            "example": "nums"
          },
          "type": {
            "type": "string",
            "description": "Parameter type",
            "example": "int[]"
          }
        }
      },
      "FunctionSignature": {
        "type": "object",
        "description": "Describes the function signature the candidate must implement",
        "properties": {
          "functionName": {
            "type": "string",
            "description": "Name of the function to implement",
            "example": "twoSum"
          },
          "params": {
            "type": "array",
            "description": "Ordered list of function parameters",
            "items": {
              "$ref": "#/components/schemas/FunctionParam"
            }
          },
          "returnType": {
            "type": "string",
            "description": "Return type of the function",
            "example": "int[]"
          }
        }
      },
...
      "QuestionResponse": {
        "type": "object",
        "description": "Question metadata, starter code templates, and visible test cases exposed to the candidate. Hidden test cases are filtered out.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique question identifier",
            "example": "two-sum"
          },
          "title": {
            "type": "string",
            "description": "Human-readable question title",
            "example": "Two Sum"
          },
          "difficulty": {
            "type": "string",
            "description": "Difficulty level",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ],
            "example": "Easy"
          },
          "description": {
            "type": "string",
            "description": "Full problem description and constraints"
          },
          "signature": {
            "$ref": "#/components/schemas/FunctionSignature",
            "description": "Function signature the candidate must implement"
          },
          "sampleInput": {
            "type": "string",
            "description": "Sample input string",
            "example": 2
          },
          "sampleOutput": {
            "type": "string",
            "description": "Sample output string",
            "example": [
              0,
              1
            ]
          },
          "starterCodes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Map of language identifier to starter code template",
            "example": {
              "java": "class Solution { ... }"
            }
          },
          "testCases": {
            "$ref": "#/components/schemas/TestCasesContainer",
            "description": "Sanitized test cases container (visible only)"
          }
        }
      },
      "TestCasesContainer": {
        "type": "object",
        "description": "Container for client-safe (visible-only) test cases",
        "properties": {
          "visible": {
            "type": "array",
            "description": "List of visible test cases shown to the candidate",
            "items": {
              "$ref": "#/components/schemas/VisibleTestCaseDto"
            }
          }
        }
      },
      "VisibleTestCaseDto": {
        "type": "object",
        "description": "A single visible test case exposed to the candidate",
        "properties": {
          "id": {
            "type": "string",
            "description": "Test case identifier",
            "example": "tc-1"
          },
          "name": {
            "type": "string",
            "description": "Human-readable display name",
            "example": "Example 1"
          },
          "inputs": {
            "type": "object",
            "additionalProperties": {},
            "description": "Map of named input parameters"
          },
          "rawInput": {
            "type": "string",
            "description": "Unparsed raw input string",
            "example": 2
          },
          "input": {
            "type": "string",
            "description": "Standard input formatted for client-side execution",
            "example": 2
          },
          "expectedOutput": {
            "type": "string",
            "description": "Expected output string",
            "example": [
              0,
              1
            ]
          },
          "explanation": {
            "type": "string",
            "description": "Optional explanation of the test case"
          }
        }
      }
    }
  }
}